requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.10.0
streamlit>=1.15.0
plotly>=5.10.0
mlflow>=2.0.0
boto3>=1.24.0
pytest>=7.1.0
black>=22.0.0
flake8>=5.0.0
PyYAML>=6.0
python-dotenv>=0.20.0
joblib>=1.1.0
requests>=2.28.0